config-loader: move EnvFunc up a level
