+++ packages/config-loader/src/sources/index.ts
@@ -25,6 +25,7 @@ export type { RemoteConfigSourceOptions } from './RemoteConfigSource';
 export { StaticConfigSource } from './StaticConfigSource';
 export type { StaticConfigSourceOptions } from './StaticConfigSource';
 export type {
+  EnvFunc,
   ConfigSource,
   ConfigSourceData,
   ReadConfigDataOptions,
