+++ packages/config-loader/src/sources/transform/types.ts
@@ -16,8 +16,6 @@
 
 import { JsonValue } from '@backstage/types';
 
-export type EnvFunc = (name: string) => Promise<string | undefined>;
-
 export type ReadFileFunc = (path: string) => Promise<string>;
 
 export interface TransformContext {
